+++ test/pip_license_checker/core_test.clj
@@ -2,11 +2,8 @@
   (:require
    [clojure.string :as str]
    [clojure.test :refer [deftest is testing]]
-   [clj-http.client :as http]
    [pip-license-checker.core :as core]
-   [pip-license-checker.file :as file]
-   [pip-license-checker.pypi :as pypi]
-   [pip-license-checker.version :as version]))
+   [pip-license-checker.pypi :as pypi]))
 
 (def params-validate-args
   [[["--requirements"
@@ -16,9 +13,11 @@
      "-r"
      "README.md"]
     {:requirements ["resources/requirements.txt" "README.md"]
+     :external []
      :packages ["django" "aiohttp==3.7.1"]
      :options {:fail #{}
                :pre false
+               :external-csv-headers true
                :with-totals false
                :totals-only false
                :table-headers false
@@ -32,55 +31,17 @@
       (testing description
         (is (= expected (core/validate-args args)))))))
 
-(def params-get-parsed-requirements
-  [[[] [] {} "{}" [] "No input"]
-   [["aiohttp==3.7.2"]
-    ["test==3.7.2"]
-    {}
-    "{\"info\": {\"license\": \"MIT License\"}}"
-    [{:ok? true,
-      :requirement {:name "aiohttp", :version "3.7.2"},
-      :license {:name "MIT License", :desc "Permissive"}}
-     {:ok? true,
-      :requirement {:name "test", :version "3.7.2"},
-      :license {:name "MIT License", :desc "Permissive"}}]
-    "Packages and requirements"]
-   [["aiohttp==3.7.2"]
-    ["test==3.7.2"]
-    {:exclude #"aio.*"}
-    "{\"info\": {\"license\": \"MIT License\"}}"
-    [{:ok? true,
-      :requirement {:name "test", :version "3.7.2"},
-      :license {:name "MIT License", :desc "Permissive"}}]
-    "Exclude pattern"]])
-
-(deftest ^:integration ^:request
-  test-get-parsed-requirements
-  (testing "Integration testing of requirements parsing"
-    (doseq [[packages requirements options mock-body expected description]
-            params-get-parsed-requirements]
-      (testing description
-        (with-redefs
-         [http/get (constantly {:body mock-body})
-          pypi/get-releases (constantly [])
-          version/get-version (constantly "3.7.2")
-          file/get-requirement-lines (fn [_] requirements)]
-          (is
-           (= expected
-              (vec (core/get-parsed-requiements
-                    packages requirements options)))))))))
-
 (def params-filter-parsed-requirements
   [[[{:ok? true,
       :requirement {:name "aiohttp", :version "3.7.2"},
-      :license {:name "MIT License", :desc "Permissive"}}
+      :license {:name "MIT License", :type "Permissive"}}
      {:ok? true,
       :requirement {:name "aiostream", :version "0.1.0"},
-      :license {:name "GPLv3", :desc "Copyleft"}}]
+      :license {:name "GPLv3", :type "Copyleft"}}]
     {:fail #{"Copyleft"} :fails-only true}
     [{:ok? true,
       :requirement {:name "aiostream", :version "0.1.0"},
-      :license {:name "GPLv3", :desc "Copyleft"}}]
+      :license {:name "GPLv3", :type "Copyleft"}}]
     "Filter copyleft licenses"]
    [[]
     {:fail #{"Copyleft"} :fails-only true}
@@ -92,31 +53,31 @@
     "Filter nil sequence"]
    [[{:ok? true,
       :requirement {:name "aiohttp", :version "3.7.2"},
-      :license {:name "MIT License", :desc "Permissive"}}
+      :license {:name "MIT License", :type "Permissive"}}
      {:ok? true,
       :requirement {:name "aiostream", :version "0.1.0"},
-      :license {:name "GPLv3", :desc "Copyleft"}}]
+      :license {:name "GPLv3", :type "Copyleft"}}]
     {:fail #{"Copyleft"} :fails-only false}
     [{:ok? true,
       :requirement {:name "aiohttp", :version "3.7.2"},
-      :license {:name "MIT License", :desc "Permissive"}}
+      :license {:name "MIT License", :type "Permissive"}}
      {:ok? true,
       :requirement {:name "aiostream", :version "0.1.0"},
-      :license {:name "GPLv3", :desc "Copyleft"}}]
+      :license {:name "GPLv3", :type "Copyleft"}}]
     "fails-only flag is off, do not filter"]
    [[{:ok? true,
       :requirement {:name "aiohttp", :version "3.7.2"},
-      :license {:name "MIT License", :desc "Permissive"}}
+      :license {:name "MIT License", :type "Permissive"}}
      {:ok? true,
       :requirement {:name "aiostream", :version "0.1.0"},
-      :license {:name "GPLv3", :desc "Copyleft"}}]
+      :license {:name "GPLv3", :type "Copyleft"}}]
     {:fail #{} :fails-only true}
     [{:ok? true,
       :requirement {:name "aiohttp", :version "3.7.2"},
-      :license {:name "MIT License", :desc "Permissive"}}
+      :license {:name "MIT License", :type "Permissive"}}
      {:ok? true,
       :requirement {:name "aiostream", :version "0.1.0"},
-      :license {:name "GPLv3", :desc "Copyleft"}}]
+      :license {:name "GPLv3", :type "Copyleft"}}]
     "fail flag omitted, do not filter"]])
 
 (deftest ^:integration ^:request
@@ -131,21 +92,21 @@
   [[[] {} "Empty vector"]
    [[{:ok? true,
       :requirement {:name "aiohttp", :version "3.7.2"},
-      :license {:name "MIT License", :desc "Permissive"}}]
+      :license {:name "MIT License", :type "Permissive"}}]
     {"Permissive" 1}
     "Single type"]
    [[{:ok? true,
       :requirement {:name "aiohttp", :version "3.7.2"},
-      :license {:name "MIT License", :desc "Permissive"}}
+      :license {:name "MIT License", :type "Permissive"}}
      {:ok? true,
       :requirement {:name "aiokafka", :version "0.6.0"},
-      :license {:name "Apache Software License", :desc "Permissive"}}
+      :license {:name "Apache Software License", :type "Permissive"}}
      {:ok? true,
       :requirement {:name "Synx", :version "0.0.3"},
-      :license {:name "Other/Proprietary License", :desc "Other"}}
+      :license {:name "Other/Proprietary License", :type "Other"}}
      {:ok? true,
       :requirement {:name "aiostream", :version "0.4.2"},
-      :license {:name "GPLv3", :desc "Copyleft"}}]
+      :license {:name "GPLv3", :type "Copyleft"}}]
     {"Copyleft" 1 "Permissive" 2 "Other" 1}
     "Multiple types"]])
 
@@ -189,7 +150,7 @@
       (testing description
         (let [license-data
               {:requirement {:name package :version version}
-               :license {:name license-name :desc license-type}}
+               :license {:name license-name :type license-type}}
               actual (core/format-license license-data)]
           (is (= expected actual)))))))
 
@@ -214,13 +175,13 @@
     "No licenses"]
    [[{:ok? true,
       :requirement {:name "test", :version "3.7.2"},
-      :license {:name "MIT License", :desc "Permissive"}}]
+      :license {:name "MIT License", :type "Permissive"}}]
     {:fail #{} :with-totals false :totals-only false :table-headers false}
     (str (format core/formatter-license "test:3.7.2" "MIT License" "Permissive") "\n")
     "No headers"]
    [[{:ok? true,
       :requirement {:name "test", :version "3.7.2"},
-      :license {:name "MIT License", :desc "Permissive"}}]
+      :license {:name "MIT License", :type "Permissive"}}]
     {:fail #{} :with-totals false :totals-only false :table-headers true}
     (str/join
      [(str (format core/formatter-license "Requirement" "License Name" "License Type") "\n")
@@ -228,7 +189,7 @@
     "With headers"]
    [[{:ok? true,
       :requirement {:name "test", :version "3.7.2"},
-      :license {:name "MIT License", :desc "Permissive"}}]
+      :license {:name "MIT License", :type "Permissive"}}]
     {:fail #{} :with-totals true :totals-only false :table-headers true}
     (str/join
      [(str (format core/formatter-license "Requirement" "License Name" "License Type") "\n")
@@ -239,14 +200,14 @@
     "With totals"]
    [[{:ok? true,
       :requirement {:name "test", :version "3.7.2"},
-      :license {:name "MIT License", :desc "Permissive"}}]
+      :license {:name "MIT License", :type "Permissive"}}]
     {:fail #{} :with-totals false :totals-only true :table-headers false}
     (str/join
      [(str (format core/formatter-totals "Permissive" 1) "\n")])
     "Totals only"]
    [[{:ok? true,
       :requirement {:name "test", :version "3.7.2"},
-      :license {:name "MIT License", :desc "Permissive"}}]
+      :license {:name "MIT License", :type "Permissive"}}]
     {:fail #{"Permissive"} :with-totals false :totals-only true :table-headers false}
     (str/join
      [(str (format core/formatter-totals "Permissive" 1) "\n")
@@ -258,9 +219,9 @@
     (doseq [[mock options expected description] params-process-requirements]
       (testing description
         (with-redefs
-         [core/get-parsed-requiements (constantly mock)
+         [pypi/get-parsed-requiements (constantly mock)
           core/exit #(println (format "Exit code: %s" %))]
-          (let [actual (with-out-str (core/process-requirements [] [] options))]
+          (let [actual (with-out-str (core/process-requirements [] [] [] options))]
             (is (= expected actual))))))))
 
 (def params-options